+++ src/components/TodoItem.tsx
@@ -1,6 +1,12 @@
-import { props } from "./Todos";
+import { Todo } from "./Todos";
 
-const TodoItem = ({ item, setTodos }: props) => {
+const TodoItem = ({
+  item,
+  setTodos,
+}: {
+  item: Todo;
+  setTodos: React.Dispatch<React.SetStateAction<Todo[]>>;
+}) => {
   const todoRemove = () => {
     setTodos((prev) => {
       if (window.confirm("일정을 삭제하시겠습니까?")) {
@@ -12,7 +18,11 @@ const TodoItem = ({ item, setTodos }: props) => {
   };
 
   const todoCompleted = () => {
-    if (window.confirm("일정을 '완료' 처리하시겠습니까?")) {
+    if (
+      item.isDone
+        ? window.confirm("완료 취소 처리하시겠습니까?")
+        : window.confirm("일정을 '완료' 처리하시겠습니까?")
+    ) {
       setTodos((prev) => {
         return prev.map((todo) => {
           return todo.id === item.id ? { ...todo, isDone: !todo.isDone } : todo;
@@ -25,11 +35,15 @@ const TodoItem = ({ item, setTodos }: props) => {
 
   return (
     <div>
-      <div> {item.title}</div>
+      <div>{item.title}</div>
       <div>{item.content}</div>
       <div>{item.date}</div>
+      {item.isDone ? (
+        <button onClick={todoCompleted}>취소</button>
+      ) : (
+        <button onClick={todoCompleted}>완료</button>
+      )}
 
-      <button onClick={todoCompleted}>완료</button>
       <button onClick={todoRemove}>삭제</button>
     </div>
   );
