--- src/components/TodoItemList.tsx
@@ -1,28 +1,0 @@
-// import React from "react";
-// import TodoItem from "../components/TodoItem";
-// import { props } from "./Todos";
-
-// const TodoItemList = ({ todos }: props) => {
-//   return (
-//     <div>
-//       <div>
-//         <h1>🔥Working</h1>
-//         {todos.map((item) => {
-//           console.log("item", item);
-//           return <TodoItem key={item.id} item={item} />;
-//         })}
-//       </div>
-//       <div>
-//         <h1>🌈Done!</h1>
-//         {todos
-//           .filter((item) => item.isDone)
-//           .map((item) => {
-//             console.log("item", item);
-//             return <TodoItem key={item.id} item={item} />;
-//           })}
-//       </div>
-//     </div>
-//   );
-// };
-
-// export default TodoItemList;
