+++ src/components/Todos.ts
@@ -1,11 +1,7 @@
-export type TodoList = {
+export type Todo = {
   id: string;
   title: string;
   content: string;
   date: string;
   isDone: boolean;
 };
-export interface props {
-  item: TodoList;
-  setTodos: React.Dispatch<React.SetStateAction<TodoList[]>>;
-}
